+++ changelog.md
@@ -1,4 +1,7 @@
 ## Upcoming
+* Changed some fields for Potion effects
+  * `hide_particles` -> `show_particles`: defaults to true now
+  * Added `show_icon`: defaults to true
 * Fixed attribute modifiers erroring (failing to spawn the mob) if a modifier with the same id is already applied
 
 ## 6.1.0.0-alpha